+++ packages/orm/src/codegen/appCodeUpdater.ts
@@ -3,33 +3,38 @@ import * as path from 'path';
 import { updateMainFile } from './updateMainFile';
 import { updateTableFile } from './updateTableFile/updateTableFile';
 import { createBaseTableFile } from './createBaseTableFile';
-import { SetOptional } from 'orchid-core';
 import { QueryLogOptions } from 'pqb';
 
 export class AppCodeUpdaterError extends Error {}
 
 export type AppCodeUpdaterConfig = {
   tablePath(tableName: string): string;
-  baseTablePath: string;
-  baseTableName: string;
   mainFilePath: string;
   logger?: QueryLogOptions['logger'];
 };
 
+export type BaseTableParam = {
+  filePath: string;
+  name: string;
+};
+
 export const appCodeUpdater = ({
   tablePath,
-  baseTablePath,
-  baseTableName,
   mainFilePath,
-}: SetOptional<AppCodeUpdaterConfig, 'baseTableName'>): AppCodeUpdater => {
-  return async ({ ast, options, basePath, cache: cacheObject, logger }) => {
+}: AppCodeUpdaterConfig): AppCodeUpdater => {
+  return async ({
+    ast,
+    options,
+    basePath,
+    cache: cacheObject,
+    logger,
+    baseTable,
+  }) => {
     const params: AppCodeUpdaterConfig = {
       tablePath(name: string) {
         const file = tablePath(name);
         return resolvePath(basePath, file);
       },
-      baseTablePath: resolvePath(basePath, baseTablePath),
-      baseTableName: baseTableName || 'BaseTable',
       mainFilePath: resolvePath(basePath, mainFilePath),
       logger,
     };
@@ -42,13 +47,13 @@ export const appCodeUpdater = ({
         options,
         logger,
       ),
-      updateTableFile({ ...params, ast }),
+      updateTableFile({ ...params, ast, baseTable }),
     ];
 
     const cache = cacheObject as { createdBaseTable?: true };
     if (!cache.createdBaseTable) {
       promises.push(
-        createBaseTableFile(params).then(() => {
+        createBaseTableFile({ logger: params.logger, baseTable }).then(() => {
           cache.createdBaseTable = true;
         }),
       );