+++ packages/orm/src/codegen/updateTableFile/createTable.test.ts
@@ -17,10 +17,12 @@ const baseTablePath = resolve('baseTable.ts');
 const baseTableName = 'BaseTable';
 const log = jest.fn();
 const params = {
-  baseTablePath,
-  baseTableName,
   tablePath,
   logger: { ...console, log },
+  baseTable: {
+    filePath: baseTablePath,
+    name: baseTableName,
+  },
 };
 
 const path = tablePath('fooBar');