+++ packages/qb/myqb/CHANGELOG.md
@@ -1,5 +1,12 @@
 # myqb
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - orchid-core@0.4.0
+
 ## 0.2.2
 
 ### Patch Changes