+++ packages/qb/pqb/CHANGELOG.md
@@ -1,5 +1,12 @@
 # pqb
 
+## 0.11.3
+
+### Patch Changes
+
+- Updated dependencies
+  - orchid-core@0.4.0
+
 ## 0.11.2
 
 ### Patch Changes